browser: add a generic `invoke` method for dispatching just about anything
Add a more flexible Browser::invoke() that dispatches the passed-in callback
along with the associated data. This isn't tied to any specific Page method
that's supposed to be invoked, everything is left to the callback.

Hypothetically other dispatch methods that do relay a method call to the Page
object could switch to this approach, but this is left open for now.