AHardwareBuffer-based export handling
Instead of rendering contents into a GLSurfaceView through SurfaceTexture
and having to relay the associated Surface object to the composition process,
make gfx.View a simple SurfaceView and provide adaptable renderers that handle
AHardwareBuffer objects as produced and managed by WPEBackend-android.

The fallback renderer is used as backup, and its approach is the one that was
used with GLSurfaceView. For every exported AHardwareBuffer, the buffer gets
resampled via EGL into the ANativeWindow that's associated with the Surface of
the SurfaceView.

The ASurfaceTranscation-based renderer avoids this copy and instead constructs
a transaction which leaves it to the display system to present the
AHardwareBuffer object in a more optimal fashion.

This enables removing a lot of cruft that was necessary with GLSurfaceView.
Additionally, support for orientation changes is enabled in this work.