Allow picking Cerbero builds from a pre-existing checkout
In addition to either checking out Cerbero to perform the necessary build or
downloading built binaries, provide a way to specify an existing Cerbero
checkout that was used to prepare and build the necessary packages.

This is really helpful when iterating on changes inside Cerbero which is
otherwise potentially tricky when done on the Cerbero checkout that's
created and then constantly reset when using the --build flag.

Build directory is now called "build". If necessary, Cerbero will be cloned
inside that directory, and the produced package will be placed under build/.
When downloading cached builds, those packages too will be placed under build/.

In this new mode, path to the Cerbero checkout is specified, and the script
expects to find the necessary packages at that location and then copies them
under build/. From this point on (common for all three approaches) the headers
and binaries are extracted from the packages and laid out so that the Android
build can use them.