[wpeview] Use mangled C++ names for entrypoints
Use the C++ mangled named for the WebKit::WebProcessMain() and
WebKit::NetworkProcessMain() functions, instead of the ones with
the android_ prefix.

This will allow dropping a patch on top of vanilla WPE WebKit that added
the android_WebProcess_main() and android_NetworkProcess_main() symbols.
The only additional thing those wrappers did was sending the list of
arguments to the Android log, for debugging purposes, but that may be
re-added at a later time inside initializeNativeMain() without needing
to modify WPE WebKit.