[minibrowser] Apply system bar insets to main view
Add an insets listener to apply the system bar insets to the browser
main view, ensuring that the main view will not cover the bars. While
at it, opt-in to edge-to-edge mode explicitly to ensure the same
behaviour also on older Android releases.

In Android 15 applications targeting SDK 35 are opted in to edge-to-edge
mode, which makes applications cover as much of the screen as possible
but may result in undesired overlaps if unhandled. For now the handling
is as simple as possible to get the existing behaviour back.

Fixes #189