[wpe] Update to WPE WebKit 2.48.2
Web Inspector resources now being shipped as a .gresource file instead
of a shared library, which needed changes in the bootstrap script and
setting the relevant environment variable. For now the .gresource file
is installed along the injected bundle, as the shared library was
previously, but this may change in the future.