[minibrowser] Apply standard insets to settings views
Apply "standard" insets (IME + system toolbars) to the Minibrowser
settings views. While at it, factor out the code into a View extension
method which gets reused by all the views that needs to have insets
applied.

Also, make sure that the insets are applied even if a View is not yet
attached. This is achieved by installing an OnAttachStateChangeListener
when the View is not yet attached, which then requests application and
removes itself afterwards.