Expose WPE WebKit and GStreamer versions to Java
This adds a new WKVersions Java class which contains static constants
for the WPE WebKit and GStreamer versions being used at run time. The
values are filled early during the JNI_OnLoad implementation, in common
code used both by the runtime (UI) and service (helper processes) entry
points, which allows using the values from any of the other Java
classes.

The first usage of the new functionality is calculating at runtime the
versioned strings used to identify unpacked assets for each of the
processes. This avoids the need to manually update those, which is easy
to forget.

In the future this could also be used for other purposes, e.g. the
MiniBrowser could show the versions in its settings activity for error
reporting and debugging purposes.